+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ### 1.0.0-alpha.2
 
-**2018-11-01**
+**2018-11-05**
 
 - 新增 顺序调整 （orderUpButtonText orderDownButtonText buttons）
 
 - 修改 slot名称 将list-remove 修改为 list-button，list-head-remove 修改为 list-head-button
+
+- 修复 block中小于minCount删除按钮不隐藏的bug
 
 ### 1.0.0-alpha.1